requirements.txt
openai
langchain
streamlit
python-dotenv
pyPDF2


-e . #install local packages